+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Fix temp directory not clearing on restart
 - Fix launching with native libraries in classpath
 - Fix logging Minecraft messages causing high cpu/ram and locking
+- Fix issue loading instances when there were none